+++ lib/init.php
@@ -379,6 +379,7 @@ function mai_load_files() {
 				'customize/site-header',
 				'customize/site-layouts',
 				'customize/typography',
+				'customize/widgets',
 				'customize/updates',
 				'customize/upsell',
 				'customize/css-cache',